ref(seer): Move Seer componenets out of gsApp, into components/seer/
