+++ static/app/components/seer/aiFeaturesAreDisabledBanner.tsx
