+++ static/app/components/seer/legacy/autofixRepoItem.tsx
