+++ static/app/components/seer/legacy/constants.ts
