+++ static/app/components/seer/legacy/selectableRepoItem.tsx
