+++ static/app/components/seer/noActiveSeerSubscriptionBanner.tsx
