+++ static/app/components/seer/projectDefaults/projectDefaultsForm.tsx
@@ -10,8 +10,7 @@ import {updateOrganization} from 'sentry/actionCreators/organizations';
 import {t, tct} from 'sentry/locale';
 import type {Organization} from 'sentry/types/organization';
 import {fetchMutation} from 'sentry/utils/queryClient';
-
-import {useCanWriteSettings} from 'getsentry/views/seerAutomation/components/useCanWriteSettings';
+import {useCanWriteSettings} from 'sentry/utils/seer/useCanWriteSettings';
 
 const schema = z.object({
   defaultAutofixAutomationTuning: z.boolean(),
