+++ static/app/components/seer/projectDetails/autofixAgent.tsx
