+++ static/app/components/seer/projectDetails/overrideHasAllValues.ts
@@ -1,6 +1,6 @@
 import type {BranchOverride} from 'sentry/components/events/autofix/types';
 
-export function isOverrideValid(override: BranchOverride) {
+export function overrideHasAllValues(override: BranchOverride) {
   return (
     override.branch_name.trim() !== '' &&
     override.tag_name.trim() !== '' &&
