+++ static/app/components/seer/repoDetails/repoDetailsForm.tsx
@@ -13,8 +13,7 @@ import type {Organization} from 'sentry/types/organization';
 import type {CodeReviewTrigger} from 'sentry/types/seer';
 import {getSeerOnboardingCheckQueryOptions} from 'sentry/utils/getSeerOnboardingCheckQueryOptions';
 import {fetchMutation, getApiQueryData, setApiQueryData} from 'sentry/utils/queryClient';
-
-import {useCanWriteSettings} from 'getsentry/views/seerAutomation/components/useCanWriteSettings';
+import {useCanWriteSettings} from 'sentry/utils/seer/useCanWriteSettings';
 
 const schema = z.object({
   enabledCodeReview: z.boolean(),
