+++ static/app/utils/seer/useCanWriteSettings.ts
