+++ static/app/views/settings/projectSeer/index.tsx
@@ -29,6 +29,8 @@ import {ExternalLink} from 'sentry/components/links/externalLink';
 import {NoAccess} from 'sentry/components/noAccess';
 import {OverrideOrDefault} from 'sentry/components/overrideOrDefault';
 import {Placeholder} from 'sentry/components/placeholder';
+import {AutofixRepositories} from 'sentry/components/seer/legacy/autofixRepositories';
+import {SEER_THRESHOLD_OPTIONS} from 'sentry/components/seer/legacy/constants';
 import {SentryDocumentTitle} from 'sentry/components/sentryDocumentTitle';
 import {t, tct} from 'sentry/locale';
 import {ProjectsStore} from 'sentry/stores/projectsStore';
@@ -43,9 +45,6 @@ import {getPricingDocsLinkForEventType} from 'sentry/views/settings/account/noti
 import {SettingsPageHeader} from 'sentry/views/settings/components/settingsPageHeader';
 import {ProjectPermissionAlert} from 'sentry/views/settings/project/projectPermissionAlert';
 import {useProjectSettingsOutlet} from 'sentry/views/settings/project/projectSettingsLayout';
-
-import {AutofixRepositories} from './autofixRepositories';
-import {SEER_THRESHOLD_OPTIONS} from './constants';
 
 const AiSetupDataConsent = OverrideOrDefault({
   overrideName: 'component:ai-setup-data-consent',
