+++ static/gsApp/views/seerAutomation/index.tsx
@@ -4,9 +4,8 @@ import {Stack} from '@sentry/scraps/layout';
 
 import {AnalyticsArea} from 'sentry/components/analyticsArea';
 import {NoAccess} from 'sentry/components/noAccess';
+import {AiFeaturesAreDisabledBanner} from 'sentry/components/seer/aiFeaturesAreDisabledBanner';
 import {useOrganization} from 'sentry/utils/useOrganization';
-
-import {AiFeaturesAreDisabledBanner} from 'getsentry/views/seerAutomation/components/aiFeaturesAreDisabledBanner';
 
 export default function SeerAutomationRoot() {
   const organization = useOrganization();
