+++ static/gsApp/views/seerAutomation/projectDetails.tsx
@@ -1,9 +1,8 @@
+import {SeerProjectDetails} from 'sentry/components/seer/projectDetails';
 import {showNewSeer} from 'sentry/utils/seer/showNewSeer';
 import {useOrganization} from 'sentry/utils/useOrganization';
 import {useProjectSettingsOutlet} from 'sentry/views/settings/project/projectSettingsLayout';
 import {ProjectSeerContainer as OldProjectDetails} from 'sentry/views/settings/projectSeer/index';
-
-import {SeerProjectDetails} from 'getsentry/views/seerAutomation/components/projectDetails';
 
 export default function SeerProjectDetailsPage() {
   const organization = useOrganization();
