+++ static/gsApp/views/seerAutomation/seerAutomation.tsx
@@ -4,6 +4,9 @@ import {Stack} from '@sentry/scraps/layout';
 
 import {NoProjectMessage} from 'sentry/components/noProjectMessage';
 import {Redirect} from 'sentry/components/redirect';
+import {NoActiveSeerSubscriptionBanner} from 'sentry/components/seer/noActiveSeerSubscriptionBanner';
+import {SeerAutomationDefault} from 'sentry/components/seer/seerAutomationDefault';
+import {SeerAutomationProjectList} from 'sentry/components/seer/seerAutomationProjectList';
 import {SentryDocumentTitle} from 'sentry/components/sentryDocumentTitle';
 import {t} from 'sentry/locale';
 import {showNewSeer} from 'sentry/utils/seer/showNewSeer';
@@ -12,9 +15,6 @@ import {useOrganization} from 'sentry/utils/useOrganization';
 import {SettingsPageHeader} from 'sentry/views/settings/components/settingsPageHeader';
 
 import {useSubscription} from 'getsentry/hooks/useSubscription';
-import {NoActiveSeerSubscriptionBanner} from 'getsentry/views/seerAutomation/components/noActiveSeerSubscriptionBanner';
-import {SeerAutomationDefault} from 'getsentry/views/seerAutomation/components/seerAutomationDefault';
-import {SeerAutomationProjectList} from 'getsentry/views/seerAutomation/components/seerAutomationProjectList';
 import {SeerConnectGitHubBanner} from 'getsentry/views/seerAutomation/components/seerConnectGitHubBanner';
 
 export default function SeerAutomation() {
